refactor: move Timer class out of TimerState interface
