--- inventory-framework-api/src/main/java/me/devnatan/inventoryframework/state/TimerState.java
@@ -1,15 +1,0 @@
-package me.devnatan.inventoryframework.state;
-
-public interface TimerState extends State<TimerState.Timer> {
-
-    interface Timer {
-
-        long initialInterval();
-
-        long currentInterval();
-
-        boolean isPaused();
-
-        boolean togglePause();
-    }
-}
