+++ inventory-framework-api/src/main/java/me/devnatan/inventoryframework/state/timer/Timer.java
@@ -1,0 +1,12 @@
+package me.devnatan.inventoryframework.state.timer;
+
+public interface Timer {
+
+	long initialInterval();
+
+	long currentInterval();
+
+	boolean isPaused();
+
+	boolean pause();
+}
