+++ inventory-framework-api/src/main/java/me/devnatan/inventoryframework/state/timer/TimerState.java
@@ -1,0 +1,6 @@
+package me.devnatan.inventoryframework.state.timer;
+
+import me.devnatan.inventoryframework.state.State;
+
+public interface TimerState extends State<Timer> {
+}
